refactor: home featured section abstracted to its own component
