+++ src/components/HomeFeaturedSection/homeFeaturedSection.scss
@@ -1,4 +1,4 @@
-@import './globals.scss';
+@import '../../app/globals.scss';
 
 .featured {
   position: relative;
